requirements.txt
alembic==1.5.4
attrs==20.3.0
autopep8==1.5.5
blinker==1.4
CacheControl==0.12.11
cachetools==5.3.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
cryptography==39.0.0
firebase-admin==6.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
httplib2==0.21.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
msgpack==1.0.4
oauth2client==4.1.3
packaging==20.9
pluggy==0.13.1
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==2.4.7
pytest==7.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
rsa==4.9
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.4
Werkzeug==1.0.1
